Introduces abstractrobot and modify panda and nLinkReacher accordingly
